Folder restructure, adding in new project.
